+++ packages/dara-core/dara/core/internal/settings.py
@@ -45,7 +45,7 @@ class Settings(BaseSettings):
     sso_jwt_algo: str = 'ES256'
     sso_verify_audience: bool = False
     sso_extra_audience: Optional[List[str]] = None
-    model_config = SettingsConfigDict(env_file='.env')
+    model_config = SettingsConfigDict(env_file='.env', extra='allow')
 
 
 def generate_env_content():