+++ packages/dara-core/changelog.md
@@ -4,9 +4,10 @@ title: Changelog
 
 ## NEXT
 
+-   Fix an issue where `Fallback` components would fail to serialize correctly
 -   Allow extra fields in the `Settings` model.
 
-## 1.16.0-alpha.1
+## 1.16.0
 
 -   Updated Dara to use `pydantic` v2.
 -   BREAKING: removed `@template` markers and the `For` component due to incompatibility with pydantic v2 and very low usage.